+++ akkurate-core/src/commonMain/kotlin/dev/nesk/akkurate/constraints/ConstraintViolation.kt
@@ -19,7 +19,7 @@ package dev.nesk.akkurate.constraints
 
 import dev.nesk.akkurate.Path
 
-public class ConstraintViolation internal constructor(public override val message: String, public override val path: Path) : ConstraintDescriptor {
+public class ConstraintViolation(public override val message: String, public override val path: Path) : ConstraintDescriptor {
     public operator fun component1(): String = message
     public operator fun component2(): Path = path