+++ akkurate-core/src/commonMain/kotlin/dev/nesk/akkurate/constraints/ConstraintViolationSet.kt
@@ -19,7 +19,7 @@ package dev.nesk.akkurate.constraints
 
 import dev.nesk.akkurate.Path
 
-public class ConstraintViolationSet internal constructor(private val messages: Set<ConstraintViolation>) : Set<ConstraintViolation> by messages {
+public class ConstraintViolationSet(private val messages: Set<ConstraintViolation>) : Set<ConstraintViolation> by messages {
     public val byPath: Map<Path, Set<ConstraintViolation>> by lazy { messages.groupBy { it.path }.mapValues { it.value.toSet() } }
 
     override fun equals(other: Any?): Boolean = messages == other